requirements.txt
absl-py==2.1.0
accelerate==0.21.0
aiofiles==23.2.1
altair==5.2.0
annotated-types==0.6.0
antlr4-python3-runtime==4.9.3
anyio==4.3.0
appdirs==1.4.4
attrs==23.2.0
beautifulsoup4==4.12.3
bitsandbytes==0.42.0
certifi==2024.2.2
charset-normalizer==3.3.2
click==8.1.7
clip==0.2.0
clip-anytorch==2.5.2
color-matcher==0.5.0
colorama==0.4.6
colour==0.1.5
contourpy==1.2.0
cycler==0.12.1
ddt==1.7.2
deepspeed==0.12.6
diffusers==0.21.4
docker-pycreds==0.4.0
docutils==0.21.2
einops==0.6.1
einops-exts==0.0.4
exceptiongroup==1.2.0
fastapi==0.110.0
ffmpy==0.3.2
filelock==3.13.1
flash-attn==2.5.6
fonttools==4.49.0
fsspec==2024.2.0
ftfy==6.2.0
gdown==5.1.0
gitdb==4.0.11
GitPython==3.1.42
gradio==4.16.0
gradio_client==0.8.1
grpcio==1.63.0
h11==0.14.0
hjson==3.1.0
httpcore==0.17.3
httpx==0.24.0
huggingface-hub==0.20.3
idna==3.6
imageio==2.34.0
imgviz==1.7.5
importlib-metadata==7.0.1
importlib_resources==6.1.2
inflect==7.2.1
Jinja2==3.1.3
joblib==1.3.2
jsonschema==4.21.1
jsonschema-specifications==2023.12.1
kiwisolver==1.4.5
kornia==0.7.2
kornia_rs==0.1.3
lazy_loader==0.3
lpips==0.1.4
Markdown==3.6
markdown-it-py==3.0.0
markdown2==2.4.13
MarkupSafe==2.1.5
matplotlib==3.8.3
mdurl==0.1.2
more-itertools==10.2.0
mpmath==1.3.0
networkx==3.2.1
ninja==1.11.1.1
numpy==1.26.4
omegaconf==2.3.0
opencv-python==4.8.0.76
orjson==3.9.15
packaging==23.2
pandas==2.2.1
peft==0.8.2
pillow==10.2.0
pillow_heif==0.16.0
protobuf==4.25.3
psutil==5.9.8
py-cpuinfo==9.0.0
pydantic==2.6.2
pydantic_core==2.16.3
pydub==0.25.1
Pygments==2.17.2
pynvml==11.5.0
pyparsing==3.1.1
PySocks==1.7.1
python-dateutil==2.8.2
python-multipart==0.0.9
pytz==2024.1
PyYAML==6.0.1
referencing==0.33.0
regex==2023.12.25
requests==2.31.0
rich==13.7.0
rpds-py==0.18.0
ruff==0.2.2
safetensors==0.4.2
scikit-image==0.22.0
scikit-learn==1.2.2
scipy==1.11.1
semantic-version==2.10.0
sentencepiece==0.1.99
sentry-sdk==1.42.0
setproctitle==1.3.3
shellingham==1.5.4
shortuuid==1.0.11
six==1.16.0
smmap==5.0.1
sniffio==1.3.1
soupsieve==2.5
starlette==0.36.3
svgwrite==1.4.3
sympy==1.12
tensorboard==2.16.2
tensorboard-data-server==0.7.2
threadpoolctl==3.3.0
tifffile==2024.2.12
timm==0.6.13
tomlkit==0.12.0
toolz==0.12.1
tqdm==4.66.2
transformers==4.37.2
tokenizers==0.15.1
triton==2.1.0
typeguard==4.2.1
typer==0.9.0
typing_extensions==4.10.0
tzdata==2024.1
urllib3==2.2.1
uvicorn==0.27.1
wandb==0.16.4
wavedrom==2.0.3.post3
wcwidth==0.2.13
websockets==11.0.3
Werkzeug==3.0.3
xformers==0.0.23.post1
zipp==3.17.0
omegaconf==2.3.0
